+++ src/main/java/com/relogiclabs/jschema/internal/builder/JArrayBuilder.java
@@ -1,7 +1,7 @@
-package com.relogiclabs.json.schema.internal.builder;
+package com.relogiclabs.jschema.internal.builder;
 
-import com.relogiclabs.json.schema.type.JArray;
-import com.relogiclabs.json.schema.type.JNode;
+import com.relogiclabs.jschema.node.JArray;
+import com.relogiclabs.jschema.node.JNode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Getter @Setter
 @Accessors(fluent = true)
-public class JArrayBuilder extends JNodeBuilder<JArrayBuilder> {
+public final class JArrayBuilder extends JNodeBuilder<JArrayBuilder> {
     private List<JNode> elements;
 
     @Override
