+++ src/main/java/com/relogiclabs/jschema/internal/builder/JBooleanBuilder.java
@@ -1,0 +1,10 @@
+package com.relogiclabs.jschema.internal.builder;
+
+import com.relogiclabs.jschema.node.JBoolean;
+
+public final class JBooleanBuilder extends JPrimitiveBuilder<Boolean> {
+    @Override
+    public JBoolean build() {
+        return JBoolean.from(this);
+    }
+}
