+++ src/main/java/com/relogiclabs/jschema/internal/builder/JDataTypeBuilder.java
@@ -1,15 +1,15 @@
-package com.relogiclabs.json.schema.internal.builder;
+package com.relogiclabs.jschema.internal.builder;
 
-import com.relogiclabs.json.schema.type.JAlias;
-import com.relogiclabs.json.schema.type.JDataType;
-import com.relogiclabs.json.schema.type.JsonType;
+import com.relogiclabs.jschema.node.JAlias;
+import com.relogiclabs.jschema.node.JDataType;
+import com.relogiclabs.jschema.node.JsonType;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
 @Getter @Setter
 @Accessors(fluent = true)
-public class JDataTypeBuilder extends JNodeBuilder<JDataTypeBuilder> {
+public final class JDataTypeBuilder extends JNodeBuilder<JDataTypeBuilder> {
     private JsonType jsonType;
     private Boolean nested;
     private JAlias alias;
