+++ src/main/java/com/relogiclabs/jschema/internal/builder/JIntegerBuilder.java
@@ -1,0 +1,10 @@
+package com.relogiclabs.jschema.internal.builder;
+
+import com.relogiclabs.jschema.node.JInteger;
+
+public final class JIntegerBuilder extends JPrimitiveBuilder<Long> {
+    @Override
+    public JInteger build() {
+        return JInteger.from(this);
+    }
+}
