+++ src/main/java/com/relogiclabs/jschema/internal/builder/JNullBuilder.java
@@ -1,0 +1,10 @@
+package com.relogiclabs.jschema.internal.builder;
+
+import com.relogiclabs.jschema.node.JNull;
+
+public final class JNullBuilder extends JNodeBuilder<JNullBuilder> {
+    @Override
+    public JNull build() {
+        return JNull.from(this);
+    }
+}
