+++ src/main/java/com/relogiclabs/jschema/internal/builder/JObjectBuilder.java
@@ -1,9 +1,9 @@
-package com.relogiclabs.json.schema.internal.builder;
+package com.relogiclabs.jschema.internal.builder;
 
-import com.relogiclabs.json.schema.collection.IndexMap;
-import com.relogiclabs.json.schema.internal.collection.IndexHashMap;
-import com.relogiclabs.json.schema.type.JObject;
-import com.relogiclabs.json.schema.type.JProperty;
+import com.relogiclabs.jschema.collection.IndexMap;
+import com.relogiclabs.jschema.internal.collection.IndexHashMap;
+import com.relogiclabs.jschema.node.JObject;
+import com.relogiclabs.jschema.node.JProperty;
 import lombok.Getter;
 import lombok.experimental.Accessors;
 
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Getter
 @Accessors(fluent = true)
-public class JObjectBuilder extends JNodeBuilder<JObjectBuilder> {
+public final class JObjectBuilder extends JNodeBuilder<JObjectBuilder> {
     private IndexMap<String, JProperty> properties;
 
     public JObjectBuilder properties(List<JProperty> properties) {
