+++ src/main/java/com/relogiclabs/jschema/internal/builder/JPropertyBuilder.java
@@ -1,14 +1,14 @@
-package com.relogiclabs.json.schema.internal.builder;
+package com.relogiclabs.jschema.internal.builder;
 
-import com.relogiclabs.json.schema.type.JNode;
-import com.relogiclabs.json.schema.type.JProperty;
+import com.relogiclabs.jschema.node.JNode;
+import com.relogiclabs.jschema.node.JProperty;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
 @Getter @Setter
 @Accessors(fluent = true)
-public class JPropertyBuilder extends JNodeBuilder<JPropertyBuilder> {
+public final class JPropertyBuilder extends JNodeBuilder<JPropertyBuilder> {
     private String key;
     private JNode value;
