+++ src/main/java/com/relogiclabs/jschema/internal/builder/JStringBuilder.java
@@ -1,0 +1,10 @@
+package com.relogiclabs.jschema.internal.builder;
+
+import com.relogiclabs.jschema.node.JString;
+
+public final class JStringBuilder extends JPrimitiveBuilder<String> {
+    @Override
+    public JString build() {
+        return JString.from(this);
+    }
+}
