+++ src/main/java/com/relogiclabs/jschema/internal/builder/JValidatorBuilder.java
@@ -1,10 +1,10 @@
-package com.relogiclabs.json.schema.internal.builder;
+package com.relogiclabs.jschema.internal.builder;
 
-import com.relogiclabs.json.schema.type.JDataType;
-import com.relogiclabs.json.schema.type.JFunction;
-import com.relogiclabs.json.schema.type.JNode;
-import com.relogiclabs.json.schema.type.JReceiver;
-import com.relogiclabs.json.schema.type.JValidator;
+import com.relogiclabs.jschema.node.JDataType;
+import com.relogiclabs.jschema.node.JFunction;
+import com.relogiclabs.jschema.node.JNode;
+import com.relogiclabs.jschema.node.JReceiver;
+import com.relogiclabs.jschema.node.JValidator;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
@@ -13,7 +13,7 @@ import java.util.List;
 
 @Getter @Setter
 @Accessors(fluent = true)
-public class JValidatorBuilder extends JNodeBuilder<JValidatorBuilder> {
+public final class JValidatorBuilder extends JNodeBuilder<JValidatorBuilder> {
     private JNode value;
     private List<JFunction> functions;
     private List<JDataType> dataTypes;
