--- src/main/java/com/relogiclabs/json/schema/internal/builder/JBooleanBuilder.java
@@ -1,10 +1,0 @@
-package com.relogiclabs.json.schema.internal.builder;
-
-import com.relogiclabs.json.schema.type.JBoolean;
-
-public class JBooleanBuilder extends JPrimitiveBuilder<Boolean> {
-    @Override
-    public JBoolean build() {
-        return JBoolean.from(this);
-    }
-}
