--- src/main/java/com/relogiclabs/json/schema/internal/builder/JFloatBuilder.java
@@ -1,10 +1,0 @@
-package com.relogiclabs.json.schema.internal.builder;
-
-import com.relogiclabs.json.schema.type.JFloat;
-
-public class JFloatBuilder extends JPrimitiveBuilder<Double> {
-    @Override
-    public JFloat build() {
-        return JFloat.from(this);
-    }
-}
