--- src/main/java/com/relogiclabs/json/schema/internal/builder/JIntegerBuilder.java
@@ -1,10 +1,0 @@
-package com.relogiclabs.json.schema.internal.builder;
-
-import com.relogiclabs.json.schema.type.JInteger;
-
-public class JIntegerBuilder extends JPrimitiveBuilder<Long> {
-    @Override
-    public JInteger build() {
-        return JInteger.from(this);
-    }
-}
