--- src/main/java/com/relogiclabs/json/schema/internal/builder/JStringBuilder.java
@@ -1,10 +1,0 @@
-package com.relogiclabs.json.schema.internal.builder;
-
-import com.relogiclabs.json.schema.type.JString;
-
-public class JStringBuilder extends JPrimitiveBuilder<String> {
-    @Override
-    public JString build() {
-        return JString.from(this);
-    }
-}
